RestGenericApiBooksApp/requirements.txt
Django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
Faker==14.2.0
requests==2.28.1
